obsync: add swift support to obsync
A single "url" doesn't make sense for a swift object store the way it does
for an S3 store or local file, so this commit explodes the interface from a
single url argument to three arguments: host, bucket, and object_prefix.

It also adds a SwiftStore and related classes.  Syncing ACLs with swift is
not yet implemented.